requirements.txt
astroid==2.3.1
bcrypt==3.1.7
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-heroku==0.1.9
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-cloud-texttospeech==0.5.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
playsound==1.2.2
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pylint==2.4.2
pytz==2019.3
requests==2.22.0
rsa==4.0
simpleaudio==1.0.2
six==1.12.0
SQLAlchemy==1.3.10
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.5
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
